Refactor tests and remove JS tests
