--- portfolio/package.json
@@ -1,5 +1,0 @@
-{
-  "scripts": {
-    "test": "mocha --recursive ./src"
-  }
-}
