+++ portfolio/src/test/java/com/google/sps/servlets/ImageServletTest.java
@@ -27,7 +27,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** */
 @RunWith(JUnit4.class)
 public class ImageServletTest {
