--- portfolio/src/test/javascript/test-comments.js
@@ -1,7 +1,0 @@
-import { equal } from 'assert';
-
-describe('Utilities', () => {
-  describe('animateElement', () => {
-    equal(true, true);
-  })
-});
